Remove unneseccary cast

+++ src/hagl.c
@@ -459,7 +459,7 @@ void hagl_scale_blit(uint16_t x0, uint16_t y0, uint16_t w, uint16_t h, bitmap_t 
         for (uint16_t x = 0; x < w; x++) {
             uint16_t px = ((x * x_ratio) >> 16);
             uint16_t py = ((y * y_ratio) >> 16);
-            color = ptr[(uint8_t)((py * source->width) + px)];
+            color = ptr[((py * source->width) + px)];
             hagl_put_pixel(x0 + x, y0 + y, color);
         }
     }
